refactor: split AlignmentModal component and extract confirm handler (#2043)
AlignmentModal's 233-line body embedded the unlink/cancel/confirm
footer (including a 34-line inline onclick closure building the
ConfirmCrossFormatLink payload) and the three-way status notice
directly in its rsx! tree. Extracts both into named sub-render
helpers in new sibling modules, matching the house style already
used for lanes.rs/choice.rs in this same split (#2016) and the
render-helper precedent from #1518: alignment_modal/footer.rs (the
action row, with handle_unlink/handle_confirm as named closures above
its rsx!) and alignment_modal/notice.rs (the stale/matched/warning
readout). The parent component body is now ~130 lines.

Pure extraction: no markup, class, testid, or behavior changes.
